05_raspberry_pi_software/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pyserial==3.5
numpy==2.2.1
opencv-python==4.10.0.84
jinja2==3.1.5
python-multipart==0.0.20
